Move serialize
