+++ packages/config/src/getConfig.ts
@@ -40,7 +40,7 @@ export function findAndEvalConfig(request: ConfigContext): ExpoConfig | null {
   if (request.configPath) {
     const config = testFileName(request.configPath);
     if (config) {
-      return config;
+      return serializeAndEvaluate(config);
     } else {
       throw new ConfigError(
         `Config with custom path ${request.configPath} couldn't be parsed.`,
@@ -52,7 +52,7 @@ export function findAndEvalConfig(request: ConfigContext): ExpoConfig | null {
   for (const configFileName of allowedConfigFileNames) {
     const configFilePath = path.join(request.projectRoot, configFileName);
     const config = testFileName(configFilePath);
-    if (config) return config;
+    if (config) return serializeAndEvaluate(config);
   }
 
   return null;
@@ -62,12 +62,9 @@ export function findAndEvalConfig(request: ConfigContext): ExpoConfig | null {
 // If they don't add support for async Webpack configs then we may need to pull support for Next.js.
 function evalConfig(configFile: string, request: ConfigContext): Partial<ExpoConfig> {
   let result: any;
-  let format: string = '';
   if (configFile.endsWith('.json5') || configFile.endsWith('.json')) {
-    format = 'json';
     result = JsonFile.read(configFile, { json5: true });
   } else {
-    format = 'js';
     result = require(configFile);
     if (result.default != null) {
       result = result.default;
@@ -110,4 +107,25 @@ function evalConfig(configFile: string, request: ConfigContext): Partial<ExpoCon
   }
 
   return result;
+}
+
+export function serializeAndEvaluate(val: any): any {
+  if (['undefined', 'string', 'boolean', 'number'].includes(typeof val)) {
+    return val;
+  } else if (typeof val === 'function') {
+    // TODO: Bacon: Should we support async methods?
+    return val();
+  } else if (Array.isArray(val)) {
+    return val.map(serializeAndEvaluate);
+  } else if (typeof val === 'object') {
+    const output: { [key: string]: any } = {};
+    for (const property in val) {
+      if (val.hasOwnProperty(property)) {
+        output[property] = serializeAndEvaluate(val[property]);
+      }
+    }
+    return output;
+  }
+
+  throw new Error(`Unhandled item type: ${typeof val}`);
 }
